pyproject.toml
[project]
name = "symbol-mutator"
version = "0.1.0"
description = "Deterministic symbol obfuscation tool for Doclint benchmarks."
requires-python = ">=3.12"
dependencies = [
    "libcst>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "jupyter>=1.0.0",
    "ruff>=0.3.0",
]
benchmark = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/symbol_mutator"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["symbol_mutator"]
